Start extracting tests into their own modules
