+++ kate/src/lib.rs
@@ -164,6 +164,9 @@ pub mod com;
 #[cfg(feature = "std")]
 pub mod gridgen;
 
+#[cfg(test)]
+mod gridgen_tests;
+
 /// Precalculate the length of padding IEC 9797 1.
 ///
 /// # NOTE
